Move base resolvers into dedicated modules
